Create PrimeNumberTestFunction_FunctionAndIfStatements.cpp
This program determines whether a given integer is a prime number. A prime number is a natural number greater than 1 that has no positive divisors other than 1 and itself. The program first asks the user to input an integer. Then, using the is_prime function, it uses If..Else statement to check whether the number is prime by testing divisibility up to the square root of the number. If no divisors are found, it concludes that the number is prime and displays the appropriate message. Otherwise, it informs the user that the number is not prime.