Create NumberToBinaryValues_RecursiveFunctionAndDataType.cpp
This C program prints the binary representation of a positive integer using a recursive function. It begins by prompting the user to input a number. If the number is zero, it directly prints 0 as a special case. Otherwise, it calls a recursive function pb() that breaks down the number by continuously dividing it by 2 and recursively calling itself to handle the more significant bits first. Once the base case is reached (when the value is 1), the recursion starts to return, and each corresponding binary digit (n % 2) is printed.